dom/imptests/html/html/dom/documents/dom-tree-accessors/test_document.body-getter-foreign-frameset.html

<!DOCTYPE html>
<title>document.body and a frameset in a foreign namespace</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-body">
<link rel="stylesheet" href="/resources/testharness.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
document.documentElement.appendChild(
  document.createElementNS("http://example.org/test", "frameset"));
</script>
<div id="log"></div>
<script>
test(function() {
  assert_equals(document.body, document.getElementsByTagName("body")[0]);
});
</script>
